requirements.txt
# Web Framework
streamlit>=1.31.0

# LLM APIs
openai>=1.12.0
google-generativeai>=0.3.0

# Math & Visualization
numpy>=1.24.0
matplotlib>=3.7.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
